requirements.txt
pytest-django
requests
pillow
sorl-thumbnail
django
djangorestframework
drf-nested-routers